deployment/requirements.txt
fastapi==0.71.0
loguru==0.5.3
numpy==1.22.0
Pillow==8.4.0
onnxruntime==1.7.0
opencv-python==4.5.5.62
uvicorn==0.17.0
python-multipart